+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ...
 
+### Added
+
+- Added new command 'Similar' for finding columns that have the same name in other datasets
+
 ### Fixed
 
 - AlterColumnType command now shows as IsImpossible when column is part of a view or table valued function